remove getIntialProps from exhibitions
